Move the set_options helper to the existing TestCase class
